% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_project_directory.R
\name{add_templates}
\alias{add_templates}
\title{Add template files}
\usage{
add_templates(package, workflow = "drake")
}
\description{
This adds the template files (word-styles-reference-01.docx, vancouver.csl,
    _drake.R, etc) to the appropriate directories.
}
