% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{calculate_ci}
\alias{calculate_ci}
\title{Compute confidence interval bounds from confidence level}
\usage{
calculate_ci(conf.level)
}
\arguments{
\item{conf.level}{See \code{\link[regtools:calculate_ci]{regtools::calculate_ci()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of
\code{\link[regtools:calculate_ci]{regtools::calculate_ci()}} and is now defunct.
}
