% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_args}
\alias{check_args}
\title{Ensure only one non-missing value is supplied in a list of arguments}
\usage{
check_args(...)
}
\arguments{
\item{...}{Arguments to be checked. Only one of these should be non-\code{NULL}.}
}
\value{
The single non-\code{NULL} value passed to \code{...}
}
\description{
Ensure only one non-missing value is supplied in a list of arguments
}
