% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-project-utils.R
\name{check_package_name}
\alias{check_package_name}
\title{Check that a package name is valid}
\usage{
check_package_name(pkgname)
}
\arguments{
\item{pkgname}{Character string}
}
\description{
A wrapper around \code{\link{valid_package_name}} that returns
an error with an informative message if the package name is not valid.
}
