% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{cmor.tools-defunct}
\alias{cmor.tools-defunct}
\title{Defunct Functions in Package \pkg{cmor.tools}}
\description{
The functions listed here are no longer part of \pkg{cmor.tools}, and have
been moved to the individual packages within the CMOR Tools ecosystem.
}
\section{Moved to \pkg{formattr}}{

\describe{
\item{\code{create_descriptive_table}}{See \code{\link[formattr:create_descriptive_table]{formattr::create_descriptive_table()}}}
\item{\code{mean_sd}, \code{n_percent}, \code{est_ci}}{See \link[formattr:output-formatting]{formattr::output-formatting}}
\item{\code{number}, \code{number_format}, \code{percent}, \code{percent_format}, \code{comma},
\code{comma_format}}{See \link[formattr:number-formatting]{formattr::number-formatting}}
}
}

\section{Moved to \pkg{CMORplots}}{

\describe{
\item{\code{scale_colour_cmor}, \code{scale_color_cmor}, \code{scale_fill_cmor}}{See \code{\link[CMORplots:scale_colour_cmor]{CMORplots::scale_colour_cmor()}}}
\item{\code{plot_cmor_colours}, \code{plot_cmor_colors}}{See \code{\link[CMORplots:plot_cmor_colours]{CMORplots::plot_cmor_colours()}}}
\item{\code{theme_cmor}}{See \code{\link[CMORplots:theme_cmor_classic]{CMORplots::theme_cmor_classic()}}}
}
}

\section{Moved to \pkg{SF6Dvalues}}{

\describe{
\item{\code{sf6d_profile}}{See \code{\link[SF6Dvalues:sf6d_profile]{SF6Dvalues::sf6d_profile()}}}
\item{\code{sf6d_utility}}{See \code{\link[SF6Dvalues:utility]{SF6Dvalues::utility()}} (and \code{\link[SF6Dvalues:sf6d_utility]{SF6Dvalues::sf6d_utility()}})}
\item{\code{sf12_scores}}{See \code{\link[SF6Dvalues:SF12_scores]{SF6Dvalues::sf12_PCS()}} and \code{\link[SF6Dvalues:SF12_scores]{SF6Dvalues::sf12_MCS()}}}
}
}

\section{Moved to \pkg{CMORprojects}}{

\describe{
\item{\code{create_research_project}}{See \code{\link[CMORprojects:create_research_project]{CMORprojects::create_research_project()}}}
\item{\code{tar_render_manuscript}}{See \code{\link[CMORprojects:tar_render_manuscript]{CMORprojects::tar_render_manuscript()}}}
}
}

\section{Moved to \pkg{regtools}}{

\describe{
\item{\code{lincom}}{See \code{\link[regtools:lincom]{regtools::lincom()}}}
\item{\code{calculate_ci}}{See \code{\link[regtools:calculate_ci]{regtools::calculate_ci()}}}
}
}

