% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{create_descriptive_table}
\alias{create_descriptive_table}
\title{Format descriptive statistics tables}
\usage{
create_descriptive_table(
  df,
  continuous,
  discrete,
  multiresponse,
  output,
  by,
  value_continuous,
  value_discrete,
  total
)
}
\arguments{
\item{df, continuous, discrete, multiresponse, output, by, value_continuous, value_discrete, total}{See \code{\link[formattr:create_descriptive_table]{formattr::create_descriptive_table()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of
\code{\link[formattr:create_descriptive_table]{formattr::create_descriptive_table()}} and is now defunct.
}
\keyword{internal}
