% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{create_research_project}
\alias{create_research_project}
\title{Create a research project using the CMOR project templates}
\usage{
create_research_project(
  path,
  workflow,
  git,
  raw_data_in_git,
  github,
  private,
  organisation,
  rstudio,
  open
)
}
\arguments{
\item{path, workflow, git, raw_data_in_git, github, private, organisation, rstudio, open}{See \code{\link[CMORprojects:create_research_project]{CMORprojects::create_research_project()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of
\code{\link[CMORprojects:create_research_project]{CMORprojects::create_research_project()}} and is now defunct.
}
