% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_names}
\alias{has_names}
\title{Does object have names?}
\usage{
has_names(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Similar to `rlang::is_named()`, but allows missing values for some names
}
