% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{lincom}
\alias{lincom}
\title{Linear combination of regression coefficients}
\usage{
lincom(model, weights, vcov.)
}
\arguments{
\item{model, weights, vcov.}{See \code{\link[regtools:lincom]{regtools::lincom()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of
\code{\link[regtools:lincom]{regtools::lincom()}} and is now defunct.
}
