% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{number-formatting}
\alias{number-formatting}
\alias{number}
\alias{number_format}
\alias{percent}
\alias{percent_format}
\alias{comma}
\alias{comma_format}
\title{Format numbers using `scales::number()`-type functions}
\usage{
number(
  x,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  trim = TRUE,
  html = TRUE,
  ...
)

number_format(
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  trim = TRUE,
  html = TRUE,
  ...
)

percent(
  x,
  accuracy = 1,
  scale = 100,
  prefix = "",
  suffix = "\%",
  big.mark = "< >",
  decimal.mark = ".",
  trim = TRUE,
  html = TRUE,
  ...
)

percent_format(
  accuracy = 1,
  scale = 100,
  prefix = "",
  suffix = "\%",
  big.mark = "< >",
  decimal.mark = ".",
  trim = TRUE,
  html = TRUE,
  ...
)

comma(
  x,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = ",",
  decimal.mark = ".",
  trim = TRUE,
  html = TRUE,
  ...
)

comma_format(
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = ",",
  decimal.mark = ".",
  trim = TRUE,
  html = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector to format}

\item{accuracy, scale, prefix, suffix, big.mark, decimal.mark, trim, ...}{As in
\code{scales::number()}}

\item{html}{Logical scalar. Whether to include formatting marks (minus
signs) as HTML strings (the default) or unicode.}
}
\description{
`r lifecycle::badge("deprecated")`

These functions are deprecated in favour of \code{\link[formattr]{nmbr}}.
}
\keyword{These}
\keyword{`scales::number()-type`}
\keyword{formatting}
\keyword{functions}
\keyword{functions.}
\keyword{internal}
\keyword{replace/extend}
\keyword{the}
