% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output-formatting}
\alias{output-formatting}
\alias{mean_sd}
\alias{n_percent}
\alias{est_ci}
\title{Results formatting for tables and reports}
\usage{
mean_sd(mean, sd, ...)

n_percent(n, proportion, ...)

est_ci(est, low, high, sep = " to ", ...)
}
\arguments{
\item{mean, sd, n, proportion, est, low, high}{Numeric vectors}

\item{...}{Passed to \code{number()}}

\item{sep}{Separator between lower and upper bounds of the CI; default is \code{' to '}.}
}
\description{
These functions are simple wrappers around \code{paste()} and \code{number()} to
format results for printing.
}
