% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{plot_cmor_colours}
\alias{plot_cmor_colours}
\alias{plot_cmor_colors}
\title{Show available CMOR colour palettes}
\usage{
plot_cmor_colours()

plot_cmor_colors()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions have been deprecated in favour of the functions in
\code{\link[CMORplots:plot_cmor_colours]{CMORplots::plot_cmor_colours()}} and are now defunct.
}
\keyword{internal}
