% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{render_manuscript}
\alias{render_manuscript}
\title{Render Rmd manuscript to docx}
\usage{
render_manuscript(
  rmd = NULL,
  out = NULL,
  reference_docx = NULL,
  csl = NULL,
  bib = NULL
)
}
\arguments{
\item{rmd}{Path to the Rmarkdown file (relative to project root)}

\item{out}{Path to the output docx file to be created}

\item{reference_docx}{Path to the reference docx (styles) file}

\item{csl}{Path to the CSL styles file}

\item{bib}{Path the the BibTex bibliography file}
}
\description{
This function renders an Rmarkdown report to Word .docx file, in the
appropriate directory and with reference, csl, and bib files as
specified.
}
