% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{sf12_scores}
\alias{sf12_scores}
\title{SF-12 component summary scores}
\usage{
sf12_scores(..., version, dimension)
}
\arguments{
\item{..., version, dimension}{See \link[SF6Dvalues:SF12_scores]{SF6Dvalues::SF12_scores}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated in favour of \code{\link[SF6Dvalues:SF12_scores]{SF6Dvalues::sf12_PCS()}} and
\code{\link[SF6Dvalues:SF12_scores]{SF6Dvalues::sf12_MCS()}} and is now defunct.
}
