% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrqol.R
\name{sf6d_profile}
\alias{sf6d_profile}
\title{SF-6D profile (SF-12 version)}
\usage{
sf6d_profile(..., version = 2, dimension = "list")
}
\arguments{
\item{...}{Vectors containing SF-12 question reponses; should be named as \code{\{Q1, Q2, ..., Q12\}},
\code{\{Q1, Q2a, ..., Q7\}}, or lower case equivalents.}
}
\description{
Construct SF-6D profile from SF-12 question-level responses
}
