% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrqol.R
\name{sf6d_utility}
\alias{sf6d_utility}
\title{SF-6D utility values}
\usage{
sf6d_utility(PF, RL, SF, PAIN, MH, VIT, values = "uk")
}
\arguments{
\item{PF, RL, SF, PAIN, MH, VIT}{SF-6D profile}

\item{values}{Not used}
}
\description{
`r lifecycle::badge("deprecated")`

This function is deprecated in favour of [SF6Dvalues::utility()] (and
    [SF6Dvalues::sf6d_utility()]).

Calculate SF-6D (SF-12) utility values using Brazier & Roberts (2004) algorithm
}
