% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{tar_render_manuscript}
\alias{tar_render_manuscript}
\title{Target to render Rmd manuscript to docx}
\usage{
tar_render_manuscript(
  name,
  path,
  output_file,
  file_reference_docx,
  file_csl,
  file_bib,
  pandoc_args,
  render_args,
  packages,
  library,
  error,
  deployment,
  priority,
  resources,
  retrieval,
  cue,
  quiet
)
}
\arguments{
\item{name, path, output_file, file_reference_docx, file_csl, file_bib, pandoc_args, render_args, packages, library, error, deployment, priority, resources, retrieval, cue, quiet}{See \code{\link[CMORprojects:tar_render_manuscript]{CMORprojects::tar_render_manuscript()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of
\code{\link[CMORprojects:tar_render_manuscript]{CMORprojects::tar_render_manuscript()}} and is now defunct.
}
