% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cmor_readme.R
\name{use_cmor_readme}
\alias{use_cmor_readme}
\title{Create a skeleton README file for CMOR research projects}
\usage{
use_cmor_readme(data)
}
\description{
This function creates a skeleton README.Rmd file with basic project
information for a CMOR research project package. \code{README.Rmd} will
be automatically added to \code{.Rbuildignore}.
}
