% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_project_directory.R
\name{use_project_directory}
\alias{use_project_directory}
\title{Create the directory structure for a CMOR research project}
\usage{
use_project_directory(
  name,
  package,
  workflow = "drake",
  git = TRUE,
  raw_data_in_git = TRUE,
  data_in_git = FALSE,
  output_in_git = FALSE
)
}
\arguments{
\item{name}{}
}
\description{
This function will create a directory structure to hold the data, code, and
outputs for a research project package, and various template files to
format output, etc.
}
