% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_proprietary_license.R
\name{use_proprietary_license}
\alias{use_proprietary_license}
\title{Attach a 'proprietary' license to a package}
\usage{
use_proprietary_license()
}
\description{
This function creates a LICENSE file indicating that the package is not open
source and not for citation, and attaches it to the current package.
}
