% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_research_project.R
\name{use_research_project}
\alias{use_research_project}
\title{Create an R package to hold a research project}
\usage{
use_research_project(path, package, fields = NULL)
}
\arguments{
\item{path}{Path to the new package. The last component of the path will
be used as the package name, and must follow standard R package naming
conventions.}
}
\description{
This function wraps usethis::create_package() to create a new R package to
    hold the data, code, and output for a research project.
}
