% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-project-utils.R
\name{valid_package_name}
\alias{valid_package_name}
\title{Check that a package name is valid}
\usage{
valid_package_name(pkgname)
}
\arguments{
\item{pkgname}{A character string}
}
\description{
Checks that a package name meets R package name requirements: it
    consists only of letters, numbers, and periods, starts with a letter, and
    does not end with a period.
}
