% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_research_project.R
\name{create_research_project}
\alias{create_research_project}
\title{Create a research project using the CMOR project templates}
\usage{
create_research_project(
  path,
  workflow = "targets",
  git = TRUE,
  raw_data_in_git = TRUE,
  github = TRUE,
  private = TRUE,
  organisation = NULL,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{Path to the new project.}

\item{workflow}{Character string: \code{"targets"}, \code{"drake"} or
\code{"make"}, to create corresponding workflow template files.}

\item{git}{Logical (default = \code{TRUE}). Whether to create a git
repository.}

\item{raw_data_in_git}{Logical (default = \code{TRUE}). If FALSE, data in the
\code{data/raw_data/} directory will be excluded from the git repository.}

\item{github}{Logical. If \code{TRUE} (the default), create a GitHub
repository and push the initial commit to GitHub.}

\item{private}{Logical. If \code{TRUE} (the default), a private GitHub
repository will be created; otherwise, the repository will be publicly
accessible. Ignored if \code{github = FALSE}.}

\item{organisation}{(Optional) Passed to \code{usethis::use_github()} to
create the repo under this organisation instead of the login associated
with the discovered GitHub token.}

\item{rstudio}{If \code{TRUE}, makes the new project into an RStudio
Project.}

\item{open}{If \code{TRUE}, activates the new project in a new RStudio
session.}
}
\description{
These function creates an R project, adds a basic README,
    initialises our standard research project directory structure and
    template files, and creates a git repository with an initial commit.
}
