% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{number-formatting}
\alias{number-formatting}
\alias{number}
\alias{number_format}
\alias{percent}
\alias{percent_format}
\alias{comma}
\alias{comma_format}
\title{Format numbers using \code{scales::number()}-type functions}
\usage{
number(
  x,
  accuracy,
  scale,
  prefix,
  suffix,
  big.mark,
  decimal.mark,
  trim,
  html,
  ...
)

number_format(
  accuracy,
  scale,
  prefix,
  suffix,
  big.mark,
  decimal.mark,
  trim,
  html,
  ...
)

percent(
  x,
  accuracy,
  scale,
  prefix,
  suffix,
  big.mark,
  decimal.mark,
  trim,
  html,
  ...
)

percent_format(
  accuracy,
  scale,
  prefix,
  suffix,
  big.mark,
  decimal.mark,
  trim,
  html,
  ...
)

comma(
  x,
  accuracy,
  scale,
  prefix,
  suffix,
  big.mark,
  decimal.mark,
  trim,
  html,
  ...
)

comma_format(
  accuracy,
  scale,
  prefix,
  suffix,
  big.mark,
  decimal.mark,
  trim,
  html,
  ...
)
}
\arguments{
\item{x, accuracy, scale, prefix, suffix, big.mark, decimal.mark, trim, ..., html}{See \link[formattr:number-formatting]{formattr::number-formatting}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions have been deprecated in favour of the functions in
\link[formattr:number-formatting]{formattr::number-formatting} and are now defunct.
}
\keyword{internal}
