% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{number-formatting}
\alias{number-formatting}
\alias{number}
\alias{number_format}
\title{Format numbers using `scales::number()`-type functions}
\usage{
number(
  x,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  trim = TRUE,
  ...
)

number_format(
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector to format}

\item{accuracy, scale, prefix, suffix, big.mark, decimal.mark, trim, ...}{As in
`scales::number()`}
}
\description{
These functions replace/extend the `scales::number()-type` formatting
    functions.
}
