% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output-formatting}
\alias{output-formatting}
\alias{mean_sd}
\alias{n_percent}
\alias{est_ci}
\title{Results formatting for tables and reports}
\usage{
mean_sd(mean, sd, ..., .glue = TRUE)

n_percent(n, proportion, ..., .glue = TRUE)

est_ci(est, low, high, ..., .sep = " to ", .glue = TRUE)
}
\arguments{
\item{mean, sd, n, proportion, est, low, high}{Numeric vectors}

\item{...}{Passed to `number()`}

\item{.glue}{Logical. Should the ouput be \code{'glue'} vector
(default=\code{TRUE}) or a regular character vector?}

\item{.sep}{Separator between lower and upper bounds of the CI; default is
\code{' to '}.}
}
\description{
`r lifecycle::badge("deprecated")`

These functions are deprecated in favour of the functions in
    \code{\link[formattr]{output-formatting}}.
}
\keyword{These}
\keyword{`number()`}
\keyword{`paste()`}
\keyword{and}
\keyword{are}
\keyword{around}
\keyword{for}
\keyword{format}
\keyword{functions}
\keyword{internal}
\keyword{printing.}
\keyword{results}
\keyword{simple}
\keyword{to}
\keyword{wrappers}
