% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output-formatting}
\alias{output-formatting}
\alias{mean_sd}
\alias{n_percent}
\title{Results formatting for tables and reports}
\usage{
mean_sd(mean, sd, ...)

n_percent(n, proportion, ...)
}
\arguments{
\item{mean, sd, n, proportion}{Numeric vectors}

\item{...}{Passed to `number()`}
}
\description{
These functions are simple wrappers around `paste()` and `number()` to
    format results for printing.
}
