% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{output-formatting}
\alias{output-formatting}
\alias{mean_sd}
\alias{n_percent}
\alias{est_ci}
\title{Results formatting for tables and reports}
\usage{
mean_sd(mean, sd, ..., .glue)

n_percent(n, proportion, ..., .glue)

est_ci(est, low, high, ..., .sep, .glue)
}
\arguments{
\item{mean, sd, n, proportion, est, low, high, ..., .sep, .glue}{See \link[formattr:output-formatting]{formattr::output-formatting}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions have been deprecated in favour of the functions in
\link[formattr:output-formatting]{formattr::output-formatting}.
}
\keyword{internal}
