% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cmor_colours}
\alias{plot_cmor_colours}
\alias{plot_cmor_colors}
\title{Show available CMOR colour palettes}
\usage{
plot_cmor_colours(palettes)

plot_cmor_colors(palettes)
}
\arguments{
\item{palettes}{Optional character vector of palettes to include in the
plot. Default is to include all available colour palettes.}
}
\description{
Plot the currently available CMOR colour palettes.
}
