% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{scale_colour_cmor}
\alias{scale_colour_cmor}
\alias{scale_color_cmor}
\alias{scale_fill_cmor}
\title{CMOR standard colour scales}
\usage{
scale_colour_cmor(..., palette, direction, grey, aesthetics)

scale_color_cmor(..., palette, direction, grey, aesthetics)

scale_fill_cmor(..., palette, direction, grey, aesthetics)
}
\arguments{
\item{..., palette, direction, grey, aesthetics}{See \code{\link[CMORplots:scale_colour_cmor]{CMORplots::scale_colour_cmor()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions have been deprecated in favour of the functions in
\code{\link[CMORplots:scale_colour_cmor]{CMORplots::scale_colour_cmor()}} and are now defunct.
}
\keyword{internal}
