% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scale_colour_cmor}
\alias{scale_colour_cmor}
\alias{scale_color_cmor}
\alias{scale_fill_cmor}
\title{CMOR standard colour scales}
\usage{
scale_colour_cmor(
  ...,
  palette = 1,
  direction = 1,
  grey = FALSE,
  aesthetics = "colour"
)

scale_color_cmor(
  ...,
  palette = 1,
  direction = 1,
  grey = FALSE,
  aesthetics = "colour"
)

scale_fill_cmor(
  ...,
  palette = 1,
  direction = 1,
  grey = FALSE,
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Passed on to \code{\link[ggplot2]{discrete_scale}} to control
name, limits, breaks, etc.}

\item{palette}{If a string, will use that named palette; if a number, will
index into the list of available palettes.}

\item{direction}{Sets the order of colours in the scale (1 for standard, -1
for reversed).}

\item{grey}{If \code{TRUE}, a grey/black colour will always be used as the
last group colour (regardless of the number of groups in the data). If
\code{FALSE} (the default), the colours and their order is as in the
original source palette.}

\item{aesthetics}{Which aesthetic(s) should the scale work with. This can be
useful, for example, to apply the same settings to both the
\code{colour} and \code{fill} aesthetics simultaneously.}
}
\description{
`r lifecycle::badge("deprecated")`

These functions are deprecated in favour of the functions in
    \code{\link[CMORplots]{scale_colour_cmor}}.
}
\keyword{'Color}
\keyword{(https://colorbrewer2.org/#type=diverging&scheme=PuOr&n=4)}
\keyword{(https://jfly.uni-koeln.de/color/#pallet)}
\keyword{(including}
\keyword{(sequential,}
\keyword{4-class}
\keyword{All}
\keyword{CMOR}
\keyword{ColorBrewer}
\keyword{Design'}
\keyword{Only}
\keyword{PuOr}
\keyword{Several}
\keyword{Sources:}
\keyword{Universal}
\keyword{\code{Bright},}
\keyword{\code{CUD}:}
\keyword{\code{High-Contrast},}
\keyword{\code{Light}:}
\keyword{\code{Medium-Contrast},}
\keyword{\code{Muted},}
\keyword{\code{Paired}:}
\keyword{\code{Vibrant},}
\keyword{and}
\keyword{are}
\keyword{colour}
\keyword{colour-blind}
\keyword{continuous}
\keyword{currently}
\keyword{diverging,}
\keyword{for}
\keyword{friendly.}
\keyword{https://personal.sron.nl/~pault/#sec:qualitative}
\keyword{implemented}
\keyword{included}
\keyword{internal}
\keyword{outputs.}
\keyword{paired)}
\keyword{palette}
\keyword{palettes}
\keyword{planned).}
\keyword{qualitative}
\keyword{scales}
\keyword{schemes}
\keyword{standard}
