% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrqol.R
\name{sf12_scores}
\alias{sf12_scores}
\title{SF-12 component summary scores}
\usage{
sf12_scores(..., version = 2, dimension = "list")
}
\arguments{
\item{...}{Vectors containing SF-12 question reponses; should be named as \code{\{Q1, Q2, ..., Q12\}},
\code{\{Q1, Q2a, ..., Q7\}}, or lower case equivalents.}
}
\description{
Calculate SF-12 PCS and MCS component summary scores from SF-12 question-level responses
}
