% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrqol.R
\name{sf12_scores}
\alias{sf12_scores}
\title{SF-12 component summary scores}
\usage{
sf12_scores(..., version = 2, dimension = "list")
}
\arguments{
\item{...}{Vectors containing SF-12 question reponses; should be named as
\code{\{Q1, Q2, ..., Q12\}}, \code{\{Q1, Q2a, ..., Q7\}}, or lower case
equivalents.}

\item{version}{Which version (1 or 2) of the SF-12 are the data based on?}

\item{dimension}{One of \code{"PCS"} or \code{"MCS"} to return the
corresponding score, or \code{"list"} (the default) to return a list
with both elements.}
}
\description{
`r lifecycle::badge("deprecated")`

This function is deprecated in favour of [SF6Dvalues::sf12_PCS()] and
    [SF6Dvalues::sf12_MCS()].

Calculate SF-12 PCS and MCS component summary scores from SF-12
    question-level responses
}
