% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrqol.R
\name{sf6d_profile}
\alias{sf6d_profile}
\title{SF-6D profile (SF-12 version)}
\usage{
sf6d_profile(..., version = 2, dimension = "list")
}
\arguments{
\item{...}{Vectors containing SF-12 question reponses; should be named as
\code{\{Q1, Q2, ..., Q12\}}, \code{\{Q1, Q2a, ..., Q7\}}, or lower case
equivalents.}

\item{version}{Which version (1 or 2) of the SF-12 are the data based on?}

\item{dimension}{One of \code{"PF"}, \code{"RL"}, \code{"SF"},
\code{"PAIN"}, \code{"MH"}, or \code{"VIT"} to return the corresponding
scores, or \code{"list"} (the default) to return a list with all
dimension scores.}
}
\description{
`r lifecycle::badge("deprecated")`

This function is deprecated in favour of [SF6Dvalues::sf6d_profile()].

Construct SF-6D profile from SF-12 question-level responses.
}
\keyword{internal}
