% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{sf6d_profile}
\alias{sf6d_profile}
\title{SF-6D profile (SF-12 version)}
\usage{
sf6d_profile(..., version, dimension)
}
\arguments{
\item{..., version, dimension}{See \code{\link[SF6Dvalues:sf6d_profile]{SF6Dvalues::sf6d_profile()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of \code{\link[SF6Dvalues:sf6d_profile]{SF6Dvalues::sf6d_profile()}}
and is now defunct.
}
\keyword{internal}
