% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{sf6d_utility}
\alias{sf6d_utility}
\title{SF-6D utility values}
\usage{
sf6d_utility(PF, RL, SF, PAIN, MH, VIT, values)
}
\arguments{
\item{PF, RL, SF, PAIN, MH, VIT, values}{See \code{\link[SF6Dvalues:utility]{SF6Dvalues::utility()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of \code{\link[SF6Dvalues:utility]{SF6Dvalues::utility()}} (and
\code{\link[SF6Dvalues:sf6d_utility]{SF6Dvalues::sf6d_utility()}}) and is now defunct.
}
