% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarchetypes.R
\name{tar_render_manuscript}
\alias{tar_render_manuscript}
\title{Target to render Rmd manuscript to docx}
\usage{
tar_render_manuscript(
  name,
  path,
  output_file,
  file_reference_docx = NULL,
  file_csl = NULL,
  file_bib = NULL,
  pandoc_args = NULL,
  render_args = NULL,
  packages = c("rmarkdown"),
  library = targets::tar_option_get("library"),
  error = targets::tar_option_get("error"),
  deployment = "main",
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue"),
  quiet = TRUE
)
}
\arguments{
\item{name}{Symbol, name of the target}

\item{path}{Character string, file path to the R Markdown source file. Must
have length 1.}

\item{output_file}{Character string, file path to the docx output file to
create.}

\item{file_reference_docx}{(optional) Symbol, existing target containing the
file path to the reference styles file used to format the output.}

\item{file_csl}{(optional) Symbol, existing target containing the file path
to the CSL file to use to format references.}

\item{file_bib}{(optional) Symbol, existing target containing the file path
to the BibTeX file holding reference details.}

\item{pandoc_args}{(optional) Additional command-line options to pass to
pandoc. See \code{\link[rmarkdown]{word_document}()}.}

\item{render_args}{(optional) Other render arguments passed to
\code{\link[rmarkdown]{render}()}. Unlike the \code{render_arguments}
argument to \code{\link[tarchetypes]{tar_render_raw}()}, this should be a
named list of arguments, not a language object containing such a list.}

\item{packages, library, error, deployment, priority, resources, retrieval, cue, quiet}{See \code{\link[tarchetypes]{tar_render_raw}()}.}
}
\description{
Render an R Markdown report to Word .docx file in a \code{targets} pipeline.
}
