% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{theme_cmor}
\alias{theme_cmor}
\title{Standard \code{ggplot} Theme for CMOR Outputs}
\usage{
theme_cmor(base_size, base_family)
}
\arguments{
\item{base_size, base_family}{See \code{\link[CMORplots:theme_cmor_classic]{CMORplots::theme_cmor_classic()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function has been deprecated in favour of
\code{\link[CMORplots:theme_cmor_classic]{CMORplots::theme_cmor_classic()}} and is now defunct.
}
\keyword{internal}
