% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{theme_cmor}
\alias{theme_cmor}
\title{Standard `ggplot` Theme for CMOR Outputs}
\usage{
theme_cmor(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size, in pts}

\item{base_family}{Base font family}
}
\description{
`r lifecycle::badge("deprecated")`

This function is deprecated in favour of
    \code{\link[CMORplots]{theme_cmor_classic}}.

This is a preliminary draft of a standard CMOR ggplot theme (just directly
    calls `theme_classic()` for now).
}
