% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_project_directory.R
\name{use_project_directory}
\alias{use_project_directory}
\title{Create the directory structure for a CMOR research project}
\usage{
use_project_directory(
  package = FALSE,
  workflow = "targets",
  git = TRUE,
  raw_data_in_git = TRUE,
  data_in_git = FALSE,
  output_in_git = FALSE
)
}
\arguments{
\item{package}{Logical (default = \code{FALSE}). Whether the project should
be created as an R package.}

\item{workflow}{Either \code{"targets"}, \code{"drake"}, or \code{"make"}, to
create corresponding workflow template files.}

\item{git}{Logical (default = \code{TRUE}). Whether to create a git
repository.}

\item{raw_data_in_git}{Logical (default = \code{TRUE}). If FALSE, data in the
\code{data/raw_data/} directory will be excluded from the git repository.}

\item{data_in_git}{Logical. If \code{FALSE} (the default), data in the
\code{data/} directory (but not the \code{data/raw_data/} subdirectory,
unless \code{raw_data_in_git} is also set to \code{FALSE}) will be
excluded from the git repository.}

\item{output_in_git}{Logical. If \code{FALSE} (the default), data in the
\code{output/} directory will be excluded from the git repository.}
}
\description{
This function will create a directory structure to hold the data, code, and
    outputs for a research project package, and various template files to
    format output, etc.
}
