% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_research_project.R
\name{use_research_project}
\alias{use_research_project}
\title{Create an R package to hold a research project}
\usage{
use_research_project(
  path,
  package,
  fields = NULL,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{Path to the new package. The last component of the path will
be used as the package name, and must follow standard R package naming
conventions.}

\item{package}{Logical (default = \code{FALSE}). Whether the project should
be created as an R package.}

\item{fields}{Named list of fields to add to the DESCRIPTION file. See
\code{\link[usethis]{create_package}} for details.}
}
\description{
This function wraps usethis::create_package() to create a new R package to
    hold the data, code, and output for a research project.
}
